Autodocument, refactor, add cause
